refactor: #102 -auth 도메인 구조를 기능 기준으로 재정리
