--- src/features/auth/actions.ts
@@ -1,30 +1,0 @@
-"use server";
-
-import { loginSchema, signupSchema } from "./schema";
-
-export async function loginAction(_prevState: unknown, formData: FormData) {
-  const parsed = loginSchema.safeParse({
-    email: formData.get("email"),
-    password: formData.get("password"),
-  });
-
-  if (!parsed.success) {
-    return { error: parsed.error.flatten().fieldErrors };
-  }
-
-  // TODO: Supabase auth
-}
-
-export async function signupAction(_prevState: unknown, formData: FormData) {
-  const parsed = signupSchema.safeParse({
-    email: formData.get("email"),
-    password: formData.get("password"),
-    confirmPassword: formData.get("confirmPassword"),
-  });
-
-  if (!parsed.success) {
-    return { error: parsed.error.flatten().fieldErrors };
-  }
-
-  // TODO: Supabase auth
-}
