+++ src/features/auth/signup/components/SignupForm.tsx
@@ -11,12 +11,12 @@ import { Checkbox } from "@/components/ui/checkbox";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { Toast } from "@/components/ui/toast";
-import { signupFormSchema } from "@/features/auth/signup/schema";
+import { signupFormSchema } from "@/features/auth/signup/schema/schema";
 import { cn } from "@/lib/utils/cn";
 
-import { isServerValidationError } from "../lib/isServerValidationError";
+import { isServerValidationError } from "../../lib/isServerValidationError";
+import { resolveFieldName } from "../../lib/resolveFieldName";
 import { mapReasonToMessage } from "../lib/mapReasonToMessage";
-import { resolveFieldName } from "../lib/resolveFieldName";
 
 export type FormInput = z.input<typeof signupFormSchema>;
 type FormValues = z.infer<typeof signupFormSchema>;
